@@ -1,7 +1,7 @@
 {
   "name": "remotedebug-ios-webkit-adapter",
   "productName": "RemoteDebug iOS WebKit Adapter",
-  "version": "0.2.2",
+  "version": "0.4.2",
   "bin": {
     "remotedebug_ios_webkit_adapter": "./out/index.js"
   },
@@ -14,31 +14,29 @@
     "lint": "gulp lint"
   },
   "dependencies": {
-    "@types/express": "^4.0.34",
-    "@types/mocha": "^2.2.40",
-    "@types/optimist": "0.0.29",
-    "child-process-promise": "^2.2.0",
-    "debug": "^2.6.3",
-    "express": "^4.14.0",
+    "child-process-promise": "^2.2.1",
+    "debug": "^4.1.1",
+    "express": "^4.17.1",
     "optimist": "^0.6.1",
-    "request": "^2.72.0",
-    "which": "^1.2.12",
-    "ws": "^1.1.0"
+    "request": "^2.88.0",
+    "which": "^1.3.1",
+    "ws": "^7.1.2"
   },
   "devDependencies": {
-    "gulp": "^3.9.1",
-    "gulp-mocha": "^2.2.0",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-tslint": "^7.1.0",
-    "gulp-typescript": "^3.1.5",
-    "mocha": "^2.5.3",
+    "@types/express": "^4.17.1",
+    "@types/mocha": "^5.2.7",
+    "@types/optimist": "0.0.29",
+    "fancy-log": "^1.3.3",
+    "gulp": "^4.0.2",
+    "gulp-mocha": "^6.0.0",
+    "gulp-sourcemaps": "^2.6.5",
+    "gulp-tslint": "^8.1.4",
+    "gulp-typescript": "^5.0.1",
+    "mocha": "^6.2.0",
     "mock-socket": "^3.0.1",
-    "mockery": "^1.7.0",
-    "tslint": "^3.15.1",
-    "typemoq": "0.0.6",
-    "typescript": "2.2.2"
-  },
-  "optionalDependencies": {
-    "vs-libimobile": "^0.0.5"
+    "mockery": "^2.1.0",
+    "tslint": "^5.19.0",
+    "typemoq": "^2.1.0",
+    "typescript": "^3.5.3"
   }
 }
