@@ -8,7 +8,7 @@ import * as path from 'path';
 import * as fs from 'fs';
 import * as os from 'os';
 import * as WebSocket from 'ws';
-import * as which from 'which'
+import * as which from 'which';
 import { execFile } from 'child-process-promise';
 import { Logger, debug } from '../logger';
 import { Adapter } from './adapter';
@@ -18,6 +18,7 @@ import { ITarget, IIOSDeviceTarget, IIOSProxySettings } from './adapterInterface
 import { IOSProtocol } from '../protocols/ios/ios';
 import { IOS8Protocol } from '../protocols/ios/ios8';
 import { IOS9Protocol } from '../protocols/ios/ios9';
+import { IOS12Protocol } from '../protocols/ios/ios12';
 
 export class IOSAdapter extends AdapterCollection {
     private _proxySettings: IIOSProxySettings;
@@ -35,8 +36,7 @@ export class IOSAdapter extends AdapterCollection {
     }
 
     public getTargets(): Promise<ITarget[]> {
-        debug(`iOSAdapter.getTargets`)
-        Logger.log('iosAdapter.getTargets');
+        debug(`iOSAdapter.getTargets`);
 
         return new Promise((resolve) => {
             request(this._url, (error: any, response: http.IncomingMessage, body: any) => {
@@ -49,29 +49,17 @@ export class IOSAdapter extends AdapterCollection {
                 resolve(devices);
             });
         }).then((devices: IIOSDeviceTarget[]) => {
-            // Now request the device version for each device found
-            const deviceVersions: Promise<IIOSDeviceTarget>[] = [];
             devices.forEach(d => {
-
-                let getter;
                 if (d.deviceId === 'SIMULATOR') {
                     d.version = '9.3.0'; // TODO: Find a way to auto detect version. Currently hardcoding it.
-                    getter = Promise.resolve(d);
+                } else if (d.deviceOSVersion) {
+                    d.version = d.deviceOSVersion;
                 } else {
-                    getter = this.getDeviceVersion(d.deviceId).then(v => {
-                        d.version = v;
-                        return Promise.resolve(d);
-                    }).catch((err) => {
-                        // Device version detection failed, using fallback
-                        Logger.log(`error.iosAdapter.getTargets.getDeviceVersion.failed.fallback, device=${d}`);
-                        d.version = '9.3.0';
-                        return Promise.resolve(d);
-                    });
+                    debug(`error.iosAdapter.getTargets.getDeviceVersion.failed.fallback, device=${d}. Please update ios-webkit-debug-proxy to version 1.8.5`);
+                    d.version = '9.3.0';
                 }
-
-                deviceVersions.push(getter);
             });
-            return Promise.all(deviceVersions);
+            return Promise.resolve(devices);
         }).then((devices: IIOSDeviceTarget[]) => {
             // Now start up all the adapters
             devices.forEach(d => {
@@ -104,7 +92,7 @@ export class IOSAdapter extends AdapterCollection {
         const target = super.connectTo(url, wsFrom);
 
         if (!target) {
-            throw new Error(`Target not found for ${url}`)
+            throw new Error(`Target not found for ${url}`);
         }
 
         if (!this._protocolMap.has(target)) {
@@ -116,7 +104,7 @@ export class IOSAdapter extends AdapterCollection {
     }
 
     public static async getProxySettings(args: any): Promise<IIOSProxySettings | string> {
-        debug(`iOSAdapter.getProxySettings`)
+        debug(`iOSAdapter.getProxySettings`);
         let settings: IIOSProxySettings = null;
 
         // Check that the proxy exists
@@ -140,74 +128,42 @@ export class IOSAdapter extends AdapterCollection {
     }
 
     private static getProxyPath(): Promise<string> {
-        debug(`iOSAdapter.getProxyPath`)
+        debug(`iOSAdapter.getProxyPath`);
         return new Promise((resolve, reject) => {
             if (os.platform() === 'win32') {
-                const x64 = os.arch() === 'x64';
-                const proxy = x64 ? path.resolve(__dirname, process.env.USERPROFILE + '/scoop/apps/ios-webkit-debug-proxy/current/ios_webkit_debug_proxy.exe') : path.resolve(__dirname, process.env.USERPROFILE + '/AppData/Roaming/npm/node_modules/vs-libimobile/lib/ios_webkit_debug_proxy.exe');
+                const proxy = process.env.SCOOP ?
+                path.resolve(__dirname, process.env.SCOOP + '/apps/ios-webkit-debug-proxy/current/ios_webkit_debug_proxy.exe') :
+                path.resolve(__dirname, process.env.USERPROFILE + '/scoop/apps/ios-webkit-debug-proxy/current/ios_webkit_debug_proxy.exe');
                 try {
                     fs.statSync(proxy);
-                    resolve(proxy)
+                    resolve(proxy);
                 } catch (err) {
-                    let message = x64 ? `ios_webkit_debug_proxy.exe not found. Please install 'scoop install ios-webkit-debug-proxy'` : `ios_webkit_debug_proxy.exe not found. Please install 'npm install -g vs-libimobile'`;
+                    let message = `ios_webkit_debug_proxy.exe not found. Please install 'scoop install ios-webkit-debug-proxy'`;
                     reject(message);
                 }
             } else if (os.platform() === 'darwin' || os.platform() === 'linux') {
                 which('ios_webkit_debug_proxy', function (err, resolvedPath) {
                     if (err) {
-                        reject('ios_webkit_debug_proxy not found. Please install ios_webkit_debug_proxy (https://github.com/google/ios-webkit-debug-proxy)')
+                        reject('ios_webkit_debug_proxy not found. Please install ios_webkit_debug_proxy (https://github.com/google/ios-webkit-debug-proxy)');
                     } else {
-                        resolve(resolvedPath)
+                        resolve(resolvedPath);
                     }
-                })
+                });
             }
-        })
+        });
     }
-
-    private static getDeviceInfoPath(): Promise<string> {
-        debug(`iOSAdapter.getDeviceInfoPath`)
-        return new Promise((resolve, reject) => {
-            if (os.platform() === 'win32') {
-                const proxy = path.resolve(__dirname, process.env.USERPROFILE + '/AppData/Roaming/npm/node_modules/vs-libimobile/lib/ideviceinfo.exe');
-                try {
-                    fs.statSync(proxy);
-                    resolve(proxy);
-                } catch (e) {
-                    reject(`ideviceinfo not found. Please install 'npm install -g vs-libimobile'`)
-                }
-
-            } else if (os.platform() === 'darwin' || os.platform() === 'linux') {
-                which('ideviceinfo', function (err, resolvedPath) {
-                    if (err) {
-                        reject('ideviceinfo not found. Please install libimobiledevice (https://github.com/libimobiledevice/libimobiledevice)')
-                    } else {
-                        resolve(resolvedPath)
-                    }
-                })
-            }
-        })
-    }
-
-    private async getDeviceVersion(uuid: string): Promise<string> {
-        debug(`iOSAdapter.getDeviceVersion`)
-        const _iDeviceInfoPath = await IOSAdapter.getDeviceInfoPath();
-        var proc = await execFile(_iDeviceInfoPath, ['-u', `${uuid}`, '-k', 'ProductVersion'])
-
-        let deviceVersion = ''
-        if (!proc.err) {
-            deviceVersion = proc.stdout.trim();
-        }
-
-        return deviceVersion;
-    };
 
     private getProtocolFor(version: string, target: Target): IOSProtocol {
-        debug(`iOSAdapter.getProtocolFor`)
+        debug(`iOSAdapter.getProtocolFor`);
         const parts = version.split('.');
         if (parts.length > 0) {
             const major = parseInt(parts[0], 10);
             if (major <= 8) {
                 return new IOS8Protocol(target);
+            }
+            const minor = parseInt(parts[1], 10);
+            if (major > 12 || major >= 12 && minor >= 2) {
+                return new IOS12Protocol(target);
             }
         }
 
