@@ -19,6 +19,8 @@ export class Target extends EventEmitter {
     private _adapterRequestMap: Map<number, { resolve: (any) => void, reject: (any) => void }>;
     private _requestId: number;
     private _id: string;
+    private _targetBased: boolean;
+    private _targetId: string;
 
     constructor(targetId: string, data?: ITarget) {
         super();
@@ -28,6 +30,8 @@ export class Target extends EventEmitter {
         this._toolRequestMap = new Map<number, string>();
         this._adapterRequestMap = new Map<number, { resolve: (any) => void, reject: (any) => void }>();
         this._requestId = 0;
+        this._targetBased = false;
+        this._targetId = null;
 
         // Chrome currently uses id, iOS usies appId
         this._id = targetId;
@@ -35,6 +39,14 @@ export class Target extends EventEmitter {
 
     public get data(): ITarget {
         return this._data;
+    }
+
+    public set targetBased(isTargetBased: boolean) {
+        this._targetBased = isTargetBased;
+    }
+
+    public set targetId(targetId: string) {
+        this._targetId = targetId;
     }
 
     public connectTo(url: string, wsFrom: WebSocket): void {
@@ -56,7 +68,7 @@ export class Target extends EventEmitter {
             this.onMessageFromTarget(message);
         });
         this._wsTarget.on('open', () => {
-            Logger.log(`Connection established to ${url}`);
+            debug(`Connection established to ${url}`);
             this._isConnected = true;
             for (let i = 0; i < this._messageBuffer.length; i++) {
                 this.onMessageFromTools(this._messageBuffer[i]);
@@ -64,7 +76,7 @@ export class Target extends EventEmitter {
             this._messageBuffer = [];
         });
         this._wsTarget.on('close', () => {
-            Logger.log('Socket is closed');
+            debug('Socket is closed');
         });
     }
 
@@ -131,7 +143,7 @@ export class Target extends EventEmitter {
 
     private onMessageFromTools(rawMessage: string): void {
         if (!this._isConnected) {
-            Logger.log('Connection not yet open, buffering message.');
+            debug('Connection not yet open, buffering message.');
             this._messageBuffer.push(rawMessage);
             return;
         }
@@ -165,7 +177,17 @@ export class Target extends EventEmitter {
     }
 
     private onMessageFromTarget(rawMessage: string): void {
-        const msg = JSON.parse(rawMessage);
+        let msg = JSON.parse(rawMessage);
+
+        if (this._targetBased) {
+            if (!msg.method || !msg.method.match(/^Target/)) {
+                return;
+            }
+            if (msg.method === 'Target.dispatchMessageFromTarget') {
+                rawMessage = msg.params.message;
+                msg = JSON.parse(rawMessage);
+            }
+        }
 
         if ('id' in msg) {
             if (this._toolRequestMap.has(msg.id)) {
@@ -245,6 +267,22 @@ export class Target extends EventEmitter {
 
     private sendToTarget(rawMessage: string): void {
         debug(`sendToTarget.${rawMessage}`);
+        if (this._targetBased) {
+            const message = JSON.parse(rawMessage);
+            if (!message.method.match(/^Target/)) {
+                const newMessage = {
+                    id: message.id,
+                    method: 'Target.sendMessageToTarget',
+                    params: {
+                        id: message.id,
+                        message: JSON.stringify(message),
+                        targetId: this._targetId
+                    }
+                };
+                rawMessage = JSON.stringify(newMessage);
+                debug(`sendToTarget.targeted.${rawMessage}`);
+            }
+        }
 
         // Make sure the target socket can receive messages
         if (this.isSocketConnected(this._wsTarget)) {
